requirements.txt
fastapi[all]
requests
httpx
pydantic
alembic
asyncpg
sqlalchemy
pydantic-settings

python-jose
bcrypt
passlib[bcrypt]